Add toPastElection, calculation and closeSsession to ElectionService

+++ src/main/java/com/bandeira/api_eleicoes/dtos/CreateElectionDTO.java
@@ -15,6 +15,6 @@ public record CreateElectionDTO(
 
         ElectionType electionType,
 
-        Long sessions
+        Double sessions
 ) {
 }
